feat: integrate

+++ poprako-util/src/page.rs
@@ -1,5 +1,8 @@
+//! Pagination parameters for offset/limit-based queries.
+
+/// Offset and limit parameters for paginated database queries.
 #[derive(Debug, Clone, Copy)]
 pub struct Page {
-    pub offset: usize,
-    pub limit: usize,
+    pub offset: u64,
+    pub limit: u64,
 }
